feature manipulator

+++ include/json_parser.h
@@ -20,12 +20,12 @@ typedef struct json_property_t {
 } json_property_t;
 
 typedef struct json_object_t {
-  size_t length;
+  int length;
   json_property_t *pairs;
 } json_object_t;
 
 typedef struct json_array_t {
-  size_t length;
+  int length;
   json_element_t **values;
 } json_array_t;
